speed up ios e2e
